Separate the colors module and add profile information to output.
